linkedinbot/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv          
selenium
webdriver-manager
langchain-google-genai
